dataflow: Allow replicating domains
Both for high availability of reads, and because we want to be able to
run reader nodes in the same process as adapters, we'd like to be able
to run multiple copies of domains within a cluster. This commit
implements the dataflow side of the replication of domains, as an extra
dimension alongside sharding, with a couple of caveats for now:

- Because atomic distributed writes are hard, we're currently unable to
  replicate domains which contain base tables, and multiple places in
  the code return an error if we ever accidentally do
- For the sake of simplicity for now, we're preventing *changing* the
  number of replicas between domains, except for the single case where
  we fan-out from 1 replica to n replicas.

In addition to just threading through replica index basically everywhere
we used to pass a shard, there are a couple of things that're required
to get this all working:

- Replay paths (really, the SetupReplayPath domain request) now have a
  `replica_fanout` field, which states whether the source of the replay
  path is replicated fewer times than the target (because of the second
  of the above caveats, that means it's always replicated once). This is
  used when sending replay requests along that replay path, to decide
  which replica to send the replay request to - either always send it to
  replica 0, or send it to the same replica as the replica making the
  request
- Egresses and Sharders (collectively "senders") now know whether the
  domain they're sending to is either replicated the same number of
  times, or fan-out (aka the current domain is replicated once, and the
  target is replicated n times). They use this information to
  replicate all messages they send, except for pieces of replays which
  only go to the replica that originally requested the replay (to avoid
  replaying the same key to a replica twice).

This is all sufficient to maintain the various partial invariants -
since replication is always surjective up the graph, any holes in
a replica always corresponds to a hole in a replica below it in the
graph.

As of this commit, nothing actually *replicates* domains at all -
there's a hardcoded `num_replicas = 1` in scheduling.rs which will be
filled in later by a function that can decide how many times a given
domain should be replicated.

Finally, there's a new `ViewFilter::Replica` filter which can be passed
when requesting a view to specifically request a specific replica of a
view, which will be useful both in testing and when we're eventually
replicating reader domains into adapters to allow those adapters to
specifically request the replica that is running in their process.

Fixes: ENG-1365
Change-Id: I9a71dd10f4deafb8c2914885e89e9af0d53249e3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2553
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>