errors: Better differentiate domains and replicas
Previously, there were two different ReadySetError variants for
referencing a domain that doesn't exist - MigrationUnknownDomain and
NoSuchDomain, where the former had an optional shard and the latter had
a required shard. Frequently, especially in controller state, we'd
just return the latter but make up a shard index of 0 when no
particular shard was being referenced. This is a little silly - it makes
more sense to instead consistently refer to the combo of a domain index
and a shard index as a "replica", and make NoSuchReplica the error for
when you are referencing a (domain,shard) pair that isn't found, and
have an entirely distinct error for referencing a domain that isn't
found. This commit does that.

Change-Id: I59c7d25f1065dd8a76c1129c6b852ca5f350ef39
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2555
Reviewed-by: Vlad Krasnov <vlad@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI