migrate: Remove vestigial booleans from node lists
As the FIXME comment in migrate/mod.rs correctly surmised, this
boolean (which, by the way, says whether a node was newly added in a
migration) was almost completely vestigial - it wasn't used anywhere
except a single place, and was passed through in a *bunch* of places
where it was basically completely unnecessary. Instead of passing it
around, it's a lot less confusing to just pass around direct lists of
node indexes, and give augmentation the set of new nodes that we already
have in scope where it's called.

Change-Id: I88cabf8fdf21b43316f4618f18518646cb72c7a2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2556
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>