Add an aggressive eviction test and some fixes
This commit adds a new integration test, that tests aggressive
eviction of BTreeMap backed queries. As a result it exposes three
bugs with the eviction:

A) the `is_empty` method for partial map returns false even when
   len == 0 causing a panic for maps that contains no keys, but do
   contain ranges.
B) range lookups assume that non-missed ranges will still be in
   the map on subsequent lookups, even though they can be evicted
C) range lookups that miss on multiple ranges will attempt to serialize
   results of multiple lookups causing an internal error

Those bugs are fixed in this change as well.

Fixes: ENG-1372
Change-Id: Ia11c47298272d6ad91ee5f36f516b609b8066360
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2560
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>