Carry over charset from underlying MySQL column
Previously we would simply hardcode the charset per column as UTF8. This
can cause issues when we are using fallback, if the actual charset for
the column being returned is actually a different charset.

This commit ensures that we carry over the column charset from the
mysql. This should only affect instances where we go to fallback.

Change-Id: If58edb3021b7b0269f814eaa83b1742f77efbde5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2558
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>