Offload PostLookup by moving sorting to PreInsertion
The PostLookup operator is currently responsible for several
"correction" operations on the results of a reader_map lookup.
All of them are pretty expensive and require multiple allocations.

To solve this problem, this CL attempts to move the sorting work
to the new PreInsertion operator, moving the load from the read
operations that are supposed to be fast, to the write operation
that is supposed to be more expensive, but also less common.

The offloading is done as follows:

* The order by step is performed on map insertion, where the rows
  are presorted, so no sort is needed in PostLookup.
* The group by step is also performed on map insertion, taking
  precedence over order by. This allows the aggregate functions to
  operate using linear scan instead of using a HashMap.

This means that currently the only transformation that requires
allocations is the aggregates step, others only require the allocation
of a Vec<Cow<[DataType]>> as a result of the collect step, but even
this will be avoided in a future CL.

There is a "regression" for contiguous ranges as the PostLookup
operation is currently applied on each key individually, but it
will also be resolved in a follow up. The PostLookup across keys
and partially missed ranges is broken anyway.

Change-Id: I0d78b35d7a3cb785f19e84dfbb9637c1983c6963
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2577
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>