benchmarks: Record CI results & provide regression info
This enables the benchmark utility's regression functionality. It
will compare the current run against data pulled from a postgres
DB. Credentials from that are pulled from SSM (IAM changes to give
SSM access already landed, and the container has the relevant
access to RDS).

If this is run on the main branch (e.g. nightly run), this will
additionally write the results to the database.

Testing this was complicated. Creation of CloudFormation images is
currently broken (specifically, the uploading), which are used for
two of the 3 targets we benchmark against (CF + k8s, docker-compose
is fine). Additionally, the pagination benchmark is currently
panicing, so CL2568 disables it for the moment.

This CL + CL2568 were tested together, along with a bunch of
commented out pipeline configs (to work around the CF issue), in
CL2156's buildkite run,
at: https://buildkite.com/readyset/readyset-benchmarks/builds/1686

Refs: ENG-1156
Change-Id: I0af98f324f0b1500ceb128c40f5f23f5f8ccd11c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2575
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>