server: Allow statically configuring replication
Allow statically configuring a "ReplicationStrategy" for deciding how
many times a domain should be replicated. The options consist of:

- `Never`, the default: never replicate domains.
- `ReaderDomains(usize)`: Replicate domains containing reader nodes n
  times, unless they also contain a base node.
- `NonBaseDomains(usize)`: Replicate domains not containing base nodes n
  times, unless they also contain a base node.

These are also exposed as (hidden) command-line options to the
noria-server binary, primarily so that clustertest can pass them later.

Obviously statically configuring the number of replicas isn't going to
get us very far in the long run, but in the short term, especially since
we're always statically setting the number of adapter instances to run
anyway, this actually is likely to get us quite far - plus it allows us
to start running tests for replication.

Change-Id: I316207f81a28944b3d94fb249bb90031d120e41d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2579
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>