Rip out node-based reader replication and regions
When we add the ability for replication to happen at a domain-level, we
won't need the graph-level replication of reader nodes any more. Without
that graph-level replication, being able to configure things like region
and primary region won't be necessary either, since those are only used
for two things; routing requests to replicated readers, and deriving
which workers are allowed to become the reader from whether they're
running in the primary region. To preserve the behavior of the latter
but with more flexibility (to allow preventing adapters from becoming
the leader too once those are running their own workers), this adds a
new `leader_eligible` field to WorkerDescriptor, threaded through from a
command-line option passed to noria-server.

This change also includes updates to a bunch of deploy config stuff like
the helm chart and cloudformation templates, neither of which have been
thoroughly tested as of this change since those aren't being actively
worked on right now (it's possible I've broken something there, but also
I'm not sure either was even working in the first place).

Change-Id: Ifc24e76f96f34d2862621b2415464250112c9e8b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2582
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>
Reviewed-by: Vlad Krasnov <vlad@readyset.io>