dataflow: Add key_count API to node state types
This complements the existing row_count API by allowing us to
specifically request a count of keys rather than rows.

Note, however, that for persistent nodes, we currently return a row
count instead of a key count. We do this because it is non-trivial and
potentially expensive to get a key count, and we do not care as much
about key counts on persistent nodes anyway. However, since this is
potentially pretty confusing, the next commit will make this clearer by
creating a special type to represent the result of the key count and
explicitly tagging results that are actually row count estimates.

Change-Id: Ie7160b8813cac0ab6c11b0fd00e629bccf473a32
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2585
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>