scheduling: Don't put two replicas on the same worker
When scheduling domain replicas, avoid putting two replicas of the same
domain shard on the same worker twice (since that'd be pretty
pointless). Since scheduling is all localized this is pretty simple - we
just have to track a set of the domain shards that have already been
scheduled onto a worker inside the scheduler, and keep that up to date
as we're scheduling new replicas.

For the sake of testing this change, this also adds a getter method to
View to expose the list of socket addresses for the shards, which since
all readers on a particular worker listen on the same TCP socket we can
then check are inequal on two replicas of the same view.

Refs: ENG-1369
Change-Id: I088531e15c022a2a7b0a24514dde46d571dbe495
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2587
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>