Properly apply post processing to multiple resultsets
In the previous change the results for each key were properly
sorted at insertion time, this change introduces a way to
merge multiple results for multiple keys into a single results
set while efficiently applying post lookup transformations with
zero copying.

The main win here is that instead of sorting the entire result
set across many keys, we now perform a k-way merge of the k sorted
results, while applying aggregates and other transformation as
needed. This sort of optimization is especially useful for queries
of the kind 'ORDER BY x LIMIT y` as it is able to lazily evaluate
the first y elements of the sorted run, but even for fully sorted
runs it is much faster than full sort. Another win is for aggregates
that now operate on sorted group-by runs as opposed to relying
on a hashmap to group rows.

The change also changes the semantics of what a result set means,
so for a query that sends multiple keys, a single result set is now
returned, as opposed to a set for each key, because that is the
behaviour noria-client actually needs to return correct results.

Overall I am seeing in excess of 10x for this query on our
"news website" benchmark:

SELECT
        A.id,
        A.title,
        A.keywords,
        A.creation_time,
FROM
        articles A, recommendations R
WHERE
        A.id = R.article_id AND
        R.user_id BETWEEN ? AND ?
ORDER BY
        A.title, A.id
LIMIT 20

Where the range is a random range between 0 to 200 keys.

The exact result is 287608 QPS vs previous 20727 QPS in standalone
mode, with p99.99 going down from 73ms to 3.8ms.

This change also fixes 25 failing logictests

Change-Id: Iab246cb0d556810de0b8061a316ee5eae126c78e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2596
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI