Store shard replicas in a better 2-dimensional array
There are a couple of reasons that Vec<Vec<T>> is a pretty poor choice
for storing 2-dimensional data, like we need to do when talking about
shards and replicas:

* There's nothing in the type system that forces the vectors to all be
  the same length - we just have to check ourselves, and hope we don't
  forget to check at any point
* An array of pointers to arrays is pretty cache-inefficient for random
  access, as it requires two memory loads.

This commit introduces a new Array2 type in its own crate, which instead
stores data in a single contiguous vector, one row after another, and
exposes a convenient API for accessing slices of individual rows, or
looking up individual cells, doing the necessary multiplication and
bounds checks internally, then uses this new type everywhere we
previously were using Vec<Vec<T>> to store shard/replica information of
various kinds.

There are other places in the codebase where we use Vec<Vec<T>> as well,
which may want to be updated to use this too, but for now I'm only
really focusing on replication so I'm keeping the change limited.

Refs: ENG-1369
Change-Id: I8924209ce773cacc865118d4d2d6fe1390a738e2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2590
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>