server: Allow preventing readers from running on servers
Add a new `--no-readers` flag that can be passed when running a
noria_server instance, which communicates to the scheduler that that
worker should *never* run domains that contain reader nodes. The idea is
to use this in conjunction with domain-level replicated embedded
readers, to set this flag on the server instance and the `--reader-only`
flag on the adapter worker, so that replicas of each reader always get
spread out across the adapter instances.

To avoid the proliferation of opaque positional boolean arguments to
functions, this also captures the previous reader_only flag and the
volume_id for a worker into a new WorkerSchedulingConfig struct, which
is passed around everywhere we were previously passing those two fields.

All of this comes with a new clustertest which creates 2 servers, one
with --no-readers passed, and 4 queries, and checks that all the readers
for those queries are scheduled onto the same worker (this is sufficient
to test this functionality since otherwise we'd evenly balance the
readers between the two workers, and in fact this test did fail without
the flag set).

Change-Id: I6c3e699ca53aad234a47b84c15fda07a4aef0d26
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2593
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>