dataflow: Implement AddAssign for the KeyCount type
This will be useful in a later commit, where we will be merging key
counts for different shards of the same node by adding them together.

There should never be a situation where we end up adding two different
types of key counts together, since all shards of a node should be the
same node type, so we panic if we ever try to add mismatched KeyCount
types.

Change-Id: I91823edbda8307c35e257ace13c9ce813f51a40d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2594
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>