controller: Add node_key_counts function
This commit adds a new function that queries all domains for node key
counts, then collects the results, merges them into a single map of
global node indexes to key counts, and returns it.

In this commit, the function is unused, but it will be used in a
subsequent commit to get node key counts and incorporate them into the
Graphviz output.

Change-Id: I69db3890a8989ffc9cbef6f306367ef3b396b085
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2595
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>