data: Allow FromSql conversion from UUID
This was implemented, but because it wasn't in the `accepts!` list it
didn't actually allow converting, causing any query to fallback that
returned a UUID to fail when deserializing the result set. This fixes,
that, and also re-enables the test that was covering this
behavior (albeit not for JSON, since postgresql doesn't allow comparing
two json values for equality for... some reason).

Fixes: https://github.com/readysettech/readyset/issues/6
Change-Id: I16eaad0cf2bc8c251dc8c469febd32806ca0c53d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2598
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI