adapter: Add --embedded-readers flag
Add an --embedded-readers flag to the adapter, which has similar
behavior to --standalone but also sets the `reader_only` and
`cannot_become_leader` flag, to be combined with --no-readers and
--reader-replicas on the adapter to make replicated reader domains run
in each adapter process.

I've tested this manually locally by running one server and two
adapters, with `--quorum 3` and `--reader-replicas 2`, creating a couple
queries and issuing reads for those queries to each adapter, and
everything seems to be very happy!

Fixes: ENG-1369
Change-Id: I5fef7f885cf7916b509a0b1c08db580e849afee5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2600
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>