Dashboard: Add Cache Hit Rate to Workload Overview
Adds a "Cache Hit Rate" column to the "Cached" table in the Workload
Overview. The hit rate is showed as a percent. The background of the
cache hit rate cells is colored according to {(0-0.6): Red, (0.6-0.8):
Yellow, (0.8-1): Green}.

This change was made to provide users information on how warm a
particular cached query is without going to the specific query
dashboard.

This cache hit rate field is not windowed, because it is supposed to
provide intuition on the latencies, which are also not windowed. If only
the last of 10 queries was a cache hit, a windowed cache hit rate metric
might show a rate of 100%, whereas the latencies would reflect 90% cache
misses.

Change-Id: I44c344d0ac849cbe69505c57860f8b5e40304fce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2602
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Jay Bensal <jay@readyset.io>