adapter: Expose flags for experimental query support
Add command-line flags and env vars to the adapter's clap struct for the
three currently-extant experimental query features - topk, paginate, and
mixed comparisons - and pass these flags on to the noria-server builder
when running in standalone mode. Also, since mixed comparisons wasn't
exposed in the noria-server binary, expose that there too for the sake
of completness.

Duplicating option fields and conversion to noria_server::Builder like
this is a temporary measure while there's some in-flight changes to the
way we run server workers inside of adapters, but once that's all landed
my plan is to replace all of this with a single command-line options
struct for the shared fields that's `#[clap(flatten)]`ed between the two
binaries.

Refs: ENG-1391
Change-Id: I70a57974c042907fedaeba2efdd4ce01f76e4d9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2608
Reviewed-by: Corbin Hall <sif@readyset.io>
Tested-by: Buildkite CI