noria: Move the dataflow KeyCount type to the noria crate
For an upcoming change, we need to be able to access the KeyCount type
from the noria crate, and the dataflow crate is not a dependency of the
noria crate. (I considered trying to re-export the type from e.g. the
noria-server crate, but that is not a dependency of the noria crate
either.)

I'm not crazy about this solution since the type is mostly pretty
specific to dataflow stuff, but ultimately, since it looks like we've
done similar things with other dataflow types, the cleanest approach
I can see is to just move the type to noria.

Change-Id: I4093e5570b8c1e5fcec12168016b1605a16c0b3b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2607
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>