installer: Implement advanced settings workflow
Add a new step to the installer before deploying the docker-compose
deployment, which gives the user the ability to change any "advanced
settings" for the deployment - namely disabling upstream SSL cert
verification, switching between explicit and implicit caching, and
enabling or disabling the three experimental query support flags. This
is all done via a new AdvancedSettings struct on the Deployment, which
subsumes the previous `migration_mode` field on the
DockerComposeDeployment, which existed but wasn't actually settable by
the user.

Fixes: ENG-1391
Change-Id: I547f67d19c2fde1e93acd6df1baaa178cc745290
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2609
Tested-by: Buildkite CI
Reviewed-by: Nick Marino <nick@readyset.io>