LocalNodeIndex::make is not actually unsafe
The `unsafe` keyword in rust should be reserved for functions which if
used incorrectly can create memory unsafety or otherwise violate the
rules of safe rust, not for functions which if used incorrectly can
create any other sorts of bugs. There's no way of using
`LocalNodeIndex::make` to create any of those cases, so it shouldn't
really be unsafe - not to mention that "wait, why is
LocalNodeIndex::make unsafe" is a question I get asked all the time.

Change-Id: I01349e3dbcea459a205b4add6de2f60493b21cf8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2611
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>