controller: Add integration test for node_key_counts function
Since normally node_key_counts would only be called from the graphviz
code, writing this test required the addition of some RPC code to let us
call node_key_counts via a handle. The test itself is relatively
straightforward; it creates a base table and a view, checks that the key
counts are 0 for both, and then checks that the key counts go up to 1
after inserting a row.

Change-Id: Ia39a5762601f4a3414c3bd81b9ba6b214cf1a5ed
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2612
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>