dashboards: Add database_type to qps legends
The queries per second query actually returns two separate time series,
one for queries executed against upstream and one for queries executed
against readyset - then when a query goes from proxied to cache we
switch. This change updates the legends of the two QPS panels on the
specific query dashboard to call out which is which.

Change-Id: I1ef3d22a436d85a97322fbdb90d683bf0d243037
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2613
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>