dataflow: Don't store reader state in Node
Currently, the write handles for reader nodes expose a similar API to
that of node state, but they are used in a different way than other
kinds of state by being stored in the *nodes* themselves, rather than in
the domain. That has been especially annoying, since it means that even
though the writer field of a Reader node is always None when used as
part of the overall dataflow state for a deployment, its presence in the
type still prevents automatically implementing Send for it.

This commit is intended as the first step on the way to removing the
distinction between the write handle for a reader and other kinds of
node state - for now, we're just moving the write handle out of the
Reader node and into a new `NodeMap<backlog::WriteHandle>` field on the
Domain itself, passed in to the various Node processing functions by
mutable reference in the same way State is. Lots of methods on the
Reader node only actually referenced the state as a field, so have been
inlined at the callsite instead of keeping them around.

This isn't *quite* enough to remove the manual `Send` and `Sync` impls
for DataflowState, but it's close - all that's left is:

- Removing or bypassing `sharded` from Sharder
- Removing `count_sum_max` from Aggregate

Change-Id: Id7a792247d1cb783ed7a9a1a840f3ce7d2900136
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2614
Reviewed-by: Nick Marino <nick@readyset.io>
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI