nom-sql: Add SqlType::Array
Add a new Array SqlType variant which recursively wraps a Box<SqlType>,
and add support for parsing postgres's array type syntax to
`type_identifier`. Everywhere that actually pattern matches on this
returns an error, for now.

PostgreSQL arrays support an optional length inside the square brackets,
which is totally ignored - so this also ignores it, not even storing it
on the AST because I want to keep things simple. If we need to in the
future, we can always add it to the AST.

Fixes: ENG-1401
Change-Id: I09092eea71aeefa0b63af56d8e1f51be8ddce0a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2616
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>