data: Add DataType::Array
Add a new Array data type, which wraps a new Array struct which
implements all the functionality of postgres's arrays, which are
n-dimensional, always rectangular, and support customizing the start
index of any dimension. Internally, the representation is an
ndarray::ArrayD - a dynamically-dimensioned array type - and a vec of
start indices, which default to 1.

Since in practice postgresql arrays are always homogenously-typed, there
may be an opportunity at some point in the future to do some clever
representation here to avoid having to deal with slices of the full
DataType enum, but for now especially as we're not sure how frequently
arrays are actually going to be used this should be ok.

Fixes: ENG-1402
Change-Id: I5df2c65244ff9cc86b10e5b1f215e377b5d8ec4a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2617
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>