nom-sql: Break literal out into its own module
It's always felt a little odd to me that the literal type definition and
its parsers are defined in `common.rs` rather than their own module -
this is about to be exacerbated a bit by me publically exporting more
functions for parsing literals, so I figure now is as good a time as any
to break it out into its own module, along with Double, Float, and
ItemPlaceholder which all only exist inside of Literal.

At some point in the future I'd like to do the same thing for SqlType.

Change-Id: Ibe9fd28a11d8ed6495e37a360a9f4c1119f0e3da
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2619
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>