Keep count for aggregates in state
Currently aggregates will return a value for a group
that no longer exists, which is both an error, and
consumes memory. To this end this change makes the
GroupedOperation keep track of the size of each group
and not return a value for empty groups.

Fixes (:checks notes) 107 logic tests.

Fixes: ENG-1420
Change-Id: I7e9a44e4d8e3d0ae92ee82ca4c36ad9b6a9ac321
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2627
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>