data: Add postgres FromSql+ToSql impls for arrays
Implement conversion for the new DataType::Array variant to + from SQL.
There's nothing built-in to the rust-postgres library to support
anything resembling the kind of multidimensional arrays we're
implementing, so this required going down a level and calling the array
utility functions in `postgres_protocol` directly - I mostly took
advantage for the impls for ToSql and FromSql rust-postgres *does* have
for Vec<T> here.

This also all comes with a new integration test for noria-data that
tests these ToSql+FromSql impls directly - later we'll do the full
end-to-end test in the noria-psql types test.

Fixes: ENG-1404
Change-Id: I837d630fb2132996c456e9274ba06d0b8f5a6f2d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2629
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI