replicators: Group config into a struct
As we add new fields to configure the replicators (beyond just the
current url, server id, and restart timeout), it's going to get
increasingly annoying to have to thread all those fields through all the
places they need to be in order to be passed to the replicator. Instead,
this commit wraps up the fields in a new replicators::Config struct,
which both derives clap::Parser so it can be used via `#[clap(flatten)]`
in the noria server binary, and `Serialize` and `Deserialize` so it can
be used via `#[serde(flatten)]` in `noria_server::Config`.

All of this is being done in preparation for adding a new field to this
config to disable verification of the SSL certificates in the
replication db for postgres, corresponding to the same flag that's
currently exposed on noria-psql.

Change-Id: I1b4b7dbc7791398d9b96f5c493eb6972306787d0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2635
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>