Make `replicated_readers` test use the Decimal type for sum
This fixes the failing clustertest in the nightly build.
The test was broken in https://gerrit.readyset.name/c/readyset/+/2623
where the return type of `sum` was changed from Int to Decimal, and
Eq does not properly compare Int with Decimal, as this would require
making sure they are also Hash equivalent.

It also updates the openssl-src dependency to make it pass cargo deny
following RUSTSEC-2022-0032.

Change-Id: I213b84ab31c91deecc38d66b315d9f459b5026c2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2638
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>