mir: Track original field order on MirQuery
When a MirQuery is used as a SQL VIEW or subquery in another query,
previously we'd use the column list of the leaf node of that query for
the list of fields returned by that query (I'm carefully using "fields"
not "columns" here since they don't have a table name, or rather the
table name is just the name of the VIEW). That worked fine, except that
we want to be able to arbitrarily reorder the columns projected by a
query for the sake of optimizing the post-lookup operation - if we do
that, queries that do a `SELECT *` from a named VIEW still need to get
the columns in the order originally specified in the `CREATE VIEW`
statement. To facilitate that, this commit adds a new `fields` field on
`MirQuery` which contains the original field list of the query as
specified by the user, then uses that when registering the schema of the
view in `SqlIncorporator::view_schemas`.

Change-Id: I8272393e1471467aa54a7cec29ec3473ad3305d8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2639
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>