data: Allow unquoted string literals in arrays
The postgresql documentation on arrays is a little unclear about this,
but empirically it is possible for pg arrays to contain string literals
that aren't quoted, as long as those string literals don't contain
braces or commas. This commit implements parsing for those, as a
separate variant besides regular nom-sql-provided literals.

Change-Id: I7a048e85ff04f5e18d09c8103d13d780e076e371
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2641
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>