data: Don't generate numeric arrays for Ord tests
Similar to the same issue we have for DataType's ord proptests, if we
generate arrays with mixed numeric types, we don't actually get a
proper transitive ordering relation, since numeric data types don't
compare well. Similar to the same case for data types, this isn't
actually a *problem*, since the only place we definitively need a strict
ordering relation is rocksdb, and in rocksdb we never store values of
different types together. So, this commit just does the same thing that
we already do for data type and just skip generating arrays of numeric
types for the ord tests.

Change-Id: I15b162de62f20266f563edb9944ac7b48934401d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2643
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>