Default to not redacting sensitive strings
To make it easier to debug things especially in end-user environments,
and since it's likely to only really end up being an enterprise-only
feature, we've decided to disable query anonymization by default,
instead requiring it to be *enabled* via a cargo feature flag rather
than disabled. To that end, this commit replaces the
launchpad/display_literals feature flag (and all corresponding feature
flags which enable it) with a new `redact_sensitive` feature flag which
has the *inverse* behavior. Also, some dockerfiles are updated and a
couple of now-obsolete (and previously-unused) ones are removed.

Fixes: ENG-1229
Change-Id: I9b9e6491957ba4ec3f7f9f23ceaed1fa436b7908
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2645
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>