server: Integrate key counts into results for GraphViz requests
This ties together a bunch of work from prior commits to allow key
counts to be included in the GraphViz output. (For base nodes, an
estimate of the row count is given instead, as indicated by the tilde
character that prefixes the key count for that node type.)

Note that key counts are only included for requests to /graph or
/graphviz (i.e. user requests that would set the `detailed` flag to true
on the call to DataflowState::graphviz). For cases where the key counts
are not included, the graph will look the same as before.

Change-Id: I75f4c220787c8d2a370c4d3f5ea8dc78ca94e29e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2649
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>