noria: Allow CREATE TABLE statements to replace previous tables.
The main idea here is that if a CREATE TABLE is issued and the table
already exists, we replace the previous table with the new one, unless
the new CREATE TABLE statement doesn't introduce any changes, in which
case it's ignored.

Doing this will allows us to modify the replicators down the line to
fully support schema changes: the idea will be that during the
snapshotting process, and while holding the DDL lock (very important to
guarantee that no changes are introduced during this process), the
replicators retrieve all the CREATE TABLE statements from all tables and
 send them to Noria, which will then decide if any change should be
 applied.
Tables that get recreated will have their replication offsets reset,
which will guarantee that the snapshot restore the data of all modified
tables.
This will also allow us to support any ALTER TABLE statement as long as
we are able to parse them, except for RENAME[1].

This change also means that all DDL statements should be introduced to
Noria through binlog replication, and binlog replication alone. Sending
CREATE TABLE statements directly to Noria would mean getting out-of-sync
 with primary.
For future work, we should further split the mechanisms we have today to
 make changes to Noria, and process DDLs from a separate method than we
 process CACHED QUERIES and VIEWs (so that we can further limit the
 Noria clients and disallow sending DDL statements).

[1]: Since we compare CREATE TABLE statements and we store them by name,
 it won't be possible to realize that what changed was actually the
 name, and we will be creating a new table. This will have to be handled
 in a special way down the line.

Refs: ENG-1362
Change-Id: I146f8db6aadc06ba969e73f78641b60a8ac9a11e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2651
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>